% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contentTypes}
\alias{contentTypes}
\title{A data frame contains the labels and groups of the contents of metabolic reconstruction}
\format{
\describe{
  \item{label}{the contents of the metabolic reconstruction}
  \item{group}{the classfication of the labels}
  }
}
\usage{
contentTypes
}
\description{
A data frame contains the labels and groups of the contents of metabolic reconstruction
}
\keyword{datasets}
