% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conversionTypes}
\alias{conversionTypes}
\title{A data frame contains the labels and groups of the transformation of metabolic reconstruction}
\format{
A data frame contains 63 rows and 2 variables:
\describe{
  \item{label}{the transformation of the metabolic reconstruction}
  \item{group}{the classfication of the labels}
  }
}
\usage{
conversionTypes
}
\description{
A data frame contains the labels and groups of the transformation of metabolic reconstruction
}
\keyword{datasets}
