% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_conversion_tree.R
\name{draw_conversion_tree}
\alias{draw_conversion_tree}
\title{draw_conversion_tree}
\usage{
draw_conversion_tree(infomatrix, Matrix, stepTypes, conversionTypes)
}
\arguments{
\item{infomatrix}{Matrix generated using the words2steps function}

\item{Matrix}{The matrix about the step or transformation or databases and tools used in the metabolic reconstruction}

\item{stepTypes}{Grouping information for reconstruction processes}

\item{conversionTypes}{Grouping information for conversion content}
}
\value{
a plot
}
\description{
Drawing of metabolic process matrix diagram and phylogenetic tree diagram
}
\examples{
\donttest{p1 <- draw_conversion_tree(matrixProcessConversion, conversionMatrix,
   stepTypes, conversionTypes)}
}
