% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_step_tree.R
\name{draw_step_tree}
\alias{draw_step_tree}
\title{draw_step_tree}
\usage{
draw_step_tree(infomatrix, Matrix, stepTypes, contentTypes)
}
\arguments{
\item{infomatrix}{Matrix generated using the words2steps function}

\item{Matrix}{The matrix about the steps or transformation or the databases and tools of the metabolic reconstruction.}

\item{stepTypes}{Grouping information for metabolic processes}

\item{contentTypes}{Grouping information for metabolic content}
}
\value{
a plot
}
\description{
Drawing of metabolic process matrix diagram and phylogenetic tree diagram
}
\examples{
\donttest{p1 <- draw_step_tree(matrixProcess, stepsMatrix, stepTypes, contentTypes)}
}
