% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_term_matrix.R
\name{get_term_matrix}
\alias{get_term_matrix}
\title{get_term_matrix}
\usage{
get_term_matrix(data)
}
\arguments{
\item{data}{information of data text}
}
\value{
words matrix
}
\description{
get_term_matrix
}
\examples{
\donttest{wordsMatrix <- get_term_matrix(text)}
}
