% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inputTxt}
\alias{inputTxt}
\title{inputTxt}
\format{
A data frame with characters:
\describe{
  \item{V1}{the contents of the metabolic reconstruction}
}
}
\usage{
inputTxt
}
\description{
A variable containing the contents of the metabolic reconstruction:
}
\keyword{datasets}
