% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_word_to_step.R
\name{map_word_to_step}
\alias{map_word_to_step}
\title{map_word_to_step}
\usage{
map_word_to_step(wordsMatrix, stepsMatrix)
}
\arguments{
\item{wordsMatrix}{The word matrix generated by the function TermDocumentMatrix in the tm package.}

\item{stepsMatrix}{Manually constructed metabolic process matrix.}
}
\value{
Specific metabolic process matrix based on text content mapping.
}
\description{
map_word_to_step
}
\examples{
\donttest{matrixProcess <- map_word_to_step(wordsMatrix, stepsMatrix)}
}
