% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matrixProcess}
\alias{matrixProcess}
\title{A data frame produced by the function map_word_to_step}
\format{
A data frame contains 67 variables
\describe{
 \item{degree}{the number of the steps used in a article.}
 \item{SpeciesName}{-1 means the input of Species name. 1 means the output of Species name.}
 \item{TaxonID}{-1 means the input of Taxon ID. 1 means the output of Taxon ID.}
 \item{KEGG}{-1 means the input of KEGG reaction. 1 means the output of KEGG reaction.}
 \item{NCBI}{-1 means the input of NCBI gene. 1 means the output of NCBI gene.}
 \item{Uniprot}{-1 means the input of Uniprot gene. 1 means the output of Uniprot.}
 \item{MetaCyc}{-1 means the input of MetaCyc gene. 1 means the output of MetaCyc.}
 \item{PubChem}{-1 means the input of PubChem molecular formula. 1 means the output of PubChem molecular formula.}
 \item{Brenda}{identify whether input the Brenda. -1 means the input of Brenda. 1 means the output of Brenda.}
 \item{Chebi}{ -1 means the input of Chebi information. 1 means the the output of Chebi information.}
 \item{WebGCM}{-1 means the input of Web GCM. 1 means the the output of Web GCM.}
 \item{SpontaneousReaction}{-1 means the input of spontaneous reaction. 1 means the the output of spontaneous reaction.}
 \item{ExtracellularAndPeriplasmicTransportReactions}{-1 means the input of Extracellular and periplasmic transport reactions.
    1 means the the output of Extracellular and periplasmic transport reactions.}
 \item{ExchangeReaction}{-1 means the input of exchange reaction. 1 means the the output of exchange reaction.}
 \item{MissingExchangeReaction}{-1 means the input of missing exchange reaction. 1 means the the output of missing exchange reaction.}
 \item{IntracellularTransportReaction}{-1 means the input of intracellular transport reaction. 1 means the the output of intracellular transport reaction.}
 \item{Gene}{-1 means the input of gene. 1 means the the output of gene.}
 \item{Protein}{-1 means the input of protein. 1 means the the output of protein.}
 \item{Knockout}{-1 means the input of knockout. 1 means the the output of knockout.}
 \item{StoichiometricMatrix}{-1 means the input of stoichiometric matrix. 1 means the the output of stoichiometric matrix.}
 \item{ObjectiveReaction}{-1 means the input of objective reaction. 1 means the the output of objective reaction.}
 \item{Constraints}{-1 means the input of constraints. 1 means the the output of constraints.}
 \item{Secretion}{-1 means the input of secretion. 1 means the the output of secretion.}
 \item{Mutisecretion}{-1 means the input of mutisecretion. 1 means the the output of mutisecretion.}
 \item{RichMedia}{-1 means the input of rich media. 1 means the the output of rich media.}
 \item{GenomeSequence}{-1 means the input of genome sequence. 1 means the the output of genome sequence.}
 \item{ProteomeSequence}{-1 means the input of proteome sequence. 1 means the the output of proteome sequence.}
 \item{AminoAcidWeight}{-1 means the input of amino acid weight. 1 means the the output of amino acid weight.}
 \item{AminoAcidMolecularWeight}{-1 means the input of amino acid molecular weight. 1 means the the output of amino acid molecular weight.}
 \item{NucleotideWeight}{-1 means the input of nucleotide weight. 1 means the the output of nucleotide weight.}
 \item{NucleotideMolecularWeight}{-1 means the input of nucleotide molecular weight. 1 means the the output of nucleotide molecular weight.}
 \item{DryWeight}{-1 means the input of dry weight. 1 means the the output of dry weight.}
 \item{BiomassReaction}{-1 means the input of biomass reaction. 1 means the the output of biomass reaction.}
 \item{DemandReaction}{-1 means the input of demand reaction. 1 means the the output of demand reaction.}
 \item{SinkReaction}{-1 means the input of sink reaction. 1 means the the output of sink reaction.}
 \item{GapReaction}{-1 means the input of gap reaction. 1 means the the output of gap reaction.}
 \item{MinORMax}{-1 means the input of gap reaction. 1 means the the output of gap reaction.}
 \item{GeneticInformation}{-1 means the input of genetic information. 1 means the the output of genetic information.}
 \item{MetabolicFunction}{-1 means the input of metabolic function. 1 means the the output of metabolic function.}
 \item{Metabolites}{-1 means the input of metabolites. 1 means the the output of metabolites.}
 \item{BiomassMetabolites}{-1 means the input of biomass metabolites. 1 means the the output of biomass metabolites.}
 \item{MetabolicReaction}{-1 means the input of metabolic reaction. 1 means the the output of metabolic reaction.}
 \item{ReconstructionData}{-1 means the input of reconstruction data. 1 means the the output of reconstruction data.}
 \item{DataStatistics}{-1 means the input of data statistics. 1 means the the output of data statistics.}
 \item{NeutralMolecularFormula}{-1 means the input of neutral molecular formula. 1 means the the output of neutral molecular formula.}
 \item{ChargedMolecularFormula}{-1 means the input of charged molecular formula. 1 means the the output of charged molecular formula.}
 \item{ConservationOfMassAndCharge}{-1 means the input of conservation of mass and charge. 1 means the the output of conservation of mass and charge.}
 \item{GibbsFreeEnergyInformation}{-1 means the input of gibbs free energy information. 1 means the the output of gibbs free energy information.}
 \item{Mass_ChargeConservationAssessment}{-1 means the input of mass-charge conservation assessment. 1 means the the output of conservation assessment.}
 \item{CellularCompartment}{-1 means the input of cellular compartment. 1 means the the output of cellular compartment.}
 \item{SubsystemInformation}{-1 means the input of subsystem information. 1 means the the output of subsystem information.}
 \item{IdentifiersInKEGG}{-1 means the input of identifiers in KEGG. 1 means the the output of identifiers in KEGG.}
 \item{IdentifiersInMetaCyc}{-1 means the input of identifiers in MetaCyc. 1 means the the output of identifiers in MetaCyc.}
 \item{UniformIdentifier}{-1 means the input of uniform identifier. 1 means the the output of uniform identifier.}
 \item{Coefficient}{-1 means the input of coefficient. 1 means the the output of coefficient.}
 \item{ScatterPlot}{-1 means the input of scatter plot. 1 means the the output of scatter plot.}
 \item{TerminalMetabolite}{-1 means the input of terminal metabolite. 1 means the the output of terminal metabolite.}
 \item{TypeIIIPathway}{-1 means the input of Type III pathway. 1 means the the output of Type III pathway.}
 \item{NetworkGap}{-1 means the input of network gap. 1 means the the output of network gap.}
 \item{Growth}{-1 means the input of growth. 1 means the the output of growth.}
 \item{BlockReaction}{-1 means the input of block reaction. 1 means the the output of block reaction.}
 \item{MetabolicFlux}{-1 means the input of metabolic flux. 1 means the the output of metabolic flux.}
 \item{ModelPredictCorrectly}{1 means the output of model PredictCorrectly.}
 \item{ModelGrowingTooFast}{-1 means the input of model growing too fast. 1 means the the output of model growing too fast.}
 \item{SBML}{-1 means the input of SBML file. 1 means the the output of SBML file.}
 \item{Mat}{-1 means the input of Mat file. 1 means the the output of Mat file.}
 \item{Excel}{-1 means the input of Excel file. 1 means the the output of Excel file.}
}
}
\usage{
matrixProcess
}
\description{
A data frame produced by the function map_word_to_step
}
\keyword{datasets}
