% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matrixProcessConversion}
\alias{matrixProcessConversion}
\title{A data frame produced by the function map_word_to_step}
\format{
A data frame contains 64 variables
\describe{
 \item{degree}{the number of the steps used in a article.}
  \item{SpeciesName}{1 means the output of Species name.}
  \item{TaxonID}{1 means the output of Taxon ID.}
  \item{NCBI}{1 means the output of NCBI.}
  \item{Uniprot}{1 means the output of Uniprot.}
  \item{KEGG}{1 means the output of KEGG.}
  \item{MetaCyc}{1 means the output of MetaCyc.}
  \item{PubChem}{1 means the output of PubChem.}
  \item{Brenda}{1 means the output of Brenda.}
  \item{Download}{1 means the output of Download.}
  \item{GeneticInformation}{1 means the output of Genetic information.}
  \item{ProteinInformation}{1 means the output of Protein information.}
  \item{GenomeSequence}{1 means the output of Genome sequence.}
  \item{ProteinSequence}{1 means the output of Protein sequence.}
  \item{MetabolicFunctionInformation}{1 means the output of Metabolic function information.}
  \item{Metabolites}{1 means the output of Metabolites.}
  \item{Cofactor}{1 means the output of Cofactor.}
  \item{Nucleotides}{1 means the output of Nucleotides.}
  \item{AminoAcid}{1 means the output of Amino acid.}
  \item{MolecularWeight}{1 means the output of Molecular weight.}
  \item{DryWeight}{1 means the output of Dry weight.}
  \item{MetabolicReaction}{1 means the output of Metabolic reaction.}
  \item{TerminalMetabolite}{1 means the output of Terminal metabolite.}
  \item{Secretion}{1 means the output of Secretion.}
  \item{BiomassReaction}{1 means the output of Biomass reaction.}
  \item{DemandReaction}{1 means the output of Demand reaction.}
  \item{SinkReaction}{1 means the output of Sink reaction.}
  \item{GapReaction}{1 means the output of Gap reaction.}
  \item{SpontaneousReaction}{1 means the output of Spontaneous reaction.}
  \item{ExtracellularAndPeriplasmicTransportReactions}{1 means the output of Extracellular and periplasmic transport reactions reaction.}
  \item{ExchangeReaction}{1 means the output of Exchange reaction.}
  \item{IntracellularTransportReaction}{1 means the output of Intracellular transport reaction.}
  \item{ReactionFlux}{1 means the output of Reaction flux.}
  \item{GPR}{1 means the output of GPR.}
  \item{BlastComparison}{1 means the output of Blast comparison.}
  \item{Homology}{1 means the output of Homology.}
  \item{HomologousGene}{1 means the output of Homology gene.}
  \item{StoichiometricMatrix}{1 means the output of Stoichiometric matrix.}
  \item{Knockout}{1 means the output of Knockout.}
  \item{TargetReaction}{1 means the output of Target reaction.}
  \item{Restrictions}{1 means the output of Restrictions.}
  \item{GrowthConditions}{1 means the output of Growth conditions.}
  \item{MinORMax}{1 means the output of min | max.}
  \item{ReconstructionData}{1 means the output of Reconstruction data.}
  \item{FVA}{1 means the output of FVA.}
  \item{MetabolicFlux}{1 means the output of Metabolic flux.}
  \item{Statistics}{1 means the output of Metabolic Statistics.}
  \item{NeutralMolecularFormula}{1 means the output of Neutral molecular formula.}
  \item{ChargedMolecularFormula}{1 means the output of Charged molecular formula.}
  \item{LiteratureDataCollection}{1 means the output of Literature data collection.}
  \item{ConservationOfMassAndCharge}{1 means the output of Conservation of mass and charge.}
  \item{GibbsFreeEnergy}{1 means the output of Gibbs free energy.}
  \item{CellCompartmentInformation}{1 means the output of Cell compartment information.}
  \item{SubsystemInformation}{1 means the output of Subsystem information.}
  \item{MetaboliteIdentification}{1 means the output of Metabolite Identification.}
  \item{Unite}{1 means the output of Unite.}
  \item{ManualPlanning}{1 means the output of Manual planning.}
  \item{Coefficient}{1 means the output of Coefficient.}
  \item{ScatterPlot}{1 means the output of Scatter plot.}
  \item{TestReport}{1 means the output of Test Report.}
  \item{TypeIIIPath}{1 means the output of Type III path.}
  \item{SBML}{1 means the output of SBML file.}
  \item{Mat}{1 means the output of Mat file.}
  \item{Excel}{1 means the output of Excel file.}
 }
}
\usage{
matrixProcessConversion
}
\description{
A data frame produced by the function map_word_to_step
}
\keyword{datasets}
