% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matrixProcessFile}
\alias{matrixProcessFile}
\title{A data frame produced by the function map_to_word in the function vizProcess}
\format{
A data frame contains with 3 variables
\describe{
 \item{step}{the steps used in the metabolic reconstruction}
 \item{degree}{the number that steps used in the metbolic reconstructoin occur in an article}
 \item{step_ID}{the order of the steps used in the metabolic reconstruction}
 }
}
\usage{
matrixProcessFile
}
\description{
A data frame produced by the function map_to_word in the function vizProcess
}
\keyword{datasets}
