% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizProcess.R
\name{slotsFunction}
\alias{slotsFunction}
\alias{slots}
\title{slotsFunction}
\usage{
slots(variablesFile)
}
\arguments{
\item{variablesFile}{A data frame contains three different variables.}
}
\description{
slotsFunction
}
\concept{vizProcess}
