% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stepTypes}
\alias{stepTypes}
\title{A data frame contains the labels and groups of the steps of metabolic reconstruction}
\format{
A data frame contains 93 rows and 2 variables:
\describe{
  \item{label}{the steps of the metabolic reconstruction}
  \item{group}{the classfication of the labels}
  }
}
\usage{
stepTypes
}
\description{
A data frame contains the labels and groups of the steps of metabolic reconstruction
}
\keyword{datasets}
