% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizProcess.R
\docType{data}
\name{systemrecon-class}
\alias{systemrecon-class}
\alias{label}
\title{systemrecon-class}
\format{
An object of class \code{character} of length 72.
}
\usage{
label
}
\description{
systemrecon-class
}
\section{Slots}{

\describe{
\item{\code{Taxon_id}}{Taxonomy ID}

\item{\code{NCBI_gene}}{Gene information from NCBI}

\item{\code{Uniprot_gene}}{Gene information from Uniprot}

\item{\code{Genetic_information}}{Genetic information}

\item{\code{KEGG_reaction}}{Reactions in KEGG}

\item{\code{MetaCyc_reaction}}{Reactions in MetaCyc}

\item{\code{Metabolic_function}}{Metabolic function}

\item{\code{Metabolic_reaction}}{Metabolic reaction}

\item{\code{Metabolites_id}}{Metabolites ID}

\item{\code{Reconstruction_data}}{Reconstruction data}

\item{\code{Genome_sequence}}{Genome sequence}

\item{\code{Genome_sequence_reference}}{Reference of genome sequence}

\item{\code{Reconstruction_data_reference}}{Reference of reconstruction data}

\item{\code{Data_statistics_metabolites}}{Data statistics of metabolites}

\item{\code{PubChem_molecular_formula}}{Molecular formula in PubChem database}

\item{\code{Brenda_molecular_formula}}{Molecular formula in Brenda database}

\item{\code{Neutral_molecular_formula}}{Neutral molecular formula}

\item{\code{Charged_molecular_formula}}{Charged molecular formula}

\item{\code{Chebi_molecular_formula}}{Chebi molecular formula}

\item{\code{Conservation_of_mass_and_charge}}{Conservation of mass and charge}

\item{\code{Web_GCM_Gibbs}}{Web GCM Gibbs}

\item{\code{Gibbs_free_energy_information}}{Information of gibbs free energy}

\item{\code{Gene_id}}{Gene ID}

\item{\code{Protein_id}}{Protein ID}

\item{\code{Cellular_compartment}}{Information of cellualr compartment}

\item{\code{Subsystem_information}}{Subsystem information}

\item{\code{Uniform_identifier_metabolites}}{Uniform identifiers of  metabollites}

\item{\code{Spontaneous_reactions}}{Spontaneous reactions}

\item{\code{Extracellular_and_periplasmic_transport_reactions}}{Extracellular and periplasmic transport reactions}

\item{\code{Exchange_reactions}}{Exchange reactions}

\item{\code{Intracellular_transport_reactions}}{ntracellular transport reactions}

\item{\code{Amino_acid_weight}}{Amino acid weight}

\item{\code{Amino_acid_molecular_weight}}{Amino acid_molecular weight}

\item{\code{Dry_weight}}{Dry weight}

\item{\code{Amino_acid_coefficient}}{Amino acid coefficient}

\item{\code{Nucleotide_coefficient}}{Nucleotide coefficient}

\item{\code{Nucleotide_weight}}{Nucleotide weight}

\item{\code{Nucleotide_molecular_weight}}{Nucleotide molecular weight}

\item{\code{Biomass_reactions}}{Biomass reactions}

\item{\code{Demand_reactions}}{Demand reactions}

\item{\code{Sink_reactions}}{Sink reactions}

\item{\code{Scatter_plot_stoichiometric_matrix}}{Scatter plot of stoichiometric matrix}

\item{\code{Objective_reaction}}{Objective reaction}

\item{\code{Constraints}}{Constraints of the model}

\item{\code{Mass_charge_conservation_assessment}}{assessment of conservation of mass and charge}

\item{\code{Terminal_metabolites}}{Terminal metabolites}

\item{\code{Gap_reactions}}{Gap reactions}

\item{\code{Missing_exchange_reactions}}{Missing exchange reactions}

\item{\code{Type_III_pathway}}{Type III pathway}

\item{\code{Network_gaps}}{Network gaps}

\item{\code{Biomass_metabolites}}{Biomass metabolites}

\item{\code{miniaml_or_maxiaml}}{miniaml or maxiaml}

\item{\code{Metabolic_flux_value}}{Value of metabolic flux}

\item{\code{Growth}}{Prediction of the growth of model}

\item{\code{Secretion_product}}{Product of secretion}

\item{\code{Mutisecretion}}{Multisecretion}

\item{\code{Rich_media}}{Environment of rich media}

\item{\code{Block_reactions}}{Block reactions}

\item{\code{Knockout}}{Knockout the single gene or reaction}

\item{\code{Model_predict_correctly}}{Predict the model correctly}

\item{\code{Model_growing_too_fast}}{Assess whether the model grow too fast}

\item{\code{Cofactors}}{Cofactors}

\item{\code{FBA}}{Flux balance analysis}

\item{\code{GPR}}{Gene-protein reaction}

\item{\code{Output}}{Output the file}

\item{\code{Iteration}}{Iteration}

\item{\code{Test}}{Test}

\item{\code{Assessment}}{Assessment}

\item{\code{Identifiers_metabolites}}{Identifiers of metabolites}

\item{\code{Output_file}}{Output the file}

\item{\code{Elemental_balance}}{Elemental balance}

\item{\code{Biomass_metabolites_coefficient}}{Coefficients of biomass and metabolites}
}}

\concept{vizProcess}
\keyword{datasets}
