% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toolsMatrix}
\alias{toolsMatrix}
\title{A data frame contains the details of the databases and tools of the metabolic reconstruction}
\format{
\describe{ A data frame contains the 93 rows and 46 varibales.
  \item{MarkerWords}{the key words of the steps of metabolic reconstruction}
  \item{ThresholdValue}{the values filter the steps  after the mapping}
  \item{Steps}{the steps of metabolic reonstruction}
  \item{Group}{the classification of the steps}
  \item{UniProtKnowledgeBase}{1 means the output of UniProt Knowledgebase.}
  \item{NCBI_Gene}{1 means the output of NCBI Gene.}
  \item{KEGG_Genes}{1 means the output of KEGG Genes.}
  \item{KEGG_Genome}{1 means the output of KEGG Genome.}
  \item{NCBI_Protein}{1 means the output of NCBI Protein.}
  \item{KEGG_Pathway}{1 means the output of KEGG Pathway.}
  \item{KEGG_Compound}{1 means the output of KEGG Compound.}
  \item{KEGG_Reaction}{1 means the output of KEGG Reaction.}
  \item{BioCyc}{1 means the output of BioCyc.}
  \item{MetaCycCompound}{1 means the output of MetaCyc Compound.}
  \item{MetaCycReaction}{1 means the output of MetaCyc Reaction.}
  \item{KEGGREST}{1 means the output of KEGGREST.}
  \item{COBRA}{1 means the output of COBRA.}
  \item{RAVEN}{1 means the output of RAVEN.}
  \item{CarveMe}{1 means the output of CarveMe.}
  \item{AuReMe}{1 means the output of AuReMe.}
  \item{MetaDraft}{1 means the output of MetaDraft.}
  \item{ModelSEED}{1 means the output of ModelSEED.}
  \item{PathwayTools}{1 means the output of Pathway Tools.}
  \item{Merlin}{1 means the output of Merlin.}
  \item{AGORA}{1 means the output of AGORA.}
  \item{COBRApy}{1 means the output of COBRApy.}
  \item{BLAST}{1 means the output of BLAST.}
  \item{ExperimentOrLiterature}{1 means the output of Experiment or literature.}
  \item{BRENDA}{1 means the output of BRENDA.}
  \item{minval}{1 means the output of minval.}
  \item{ChEBI}{1 means the output of ChEBI.}
  \item{PubChem_Compound}{1 means the output of PubChem-compound.}
  \item{ChEMBL_Compound}{1 means the output of ChEMBL compound.}
  \item{Rhea}{1 means the output of Rhea.}
  \item{pKa_DB}{1 means the output of pKa DB.}
  \item{rsbml}{1 means the output of rsbml.}
  \item{PipelinePilot}{1 means the output of Pipeline Pilot.}
  \item{Sybil}{1 means the output of Sybil.}
  \item{BUSCA}{1 means the output of BUSCA.}
  \item{PSORT}{1 means the output of PSORT.}
  \item{PASUB}{1 means the output of PASUB.}
  \item{PubChem_Substance}{1 means the output of PubChem substance.}
  \item{STRING}{1 means the output of STRING.}
  \item{ManualPlanning}{1 means the output of Manual planning.}
  \item{CMR}{1 means the output of CMR.}
  \item{g2f}{1 means the output of g2f.}
  }
}
\usage{
toolsMatrix
}
\description{
A data frame contains the details of the databases and tools of the metabolic reconstruction
}
\keyword{datasets}
