% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toolsTypes}
\alias{toolsTypes}
\title{A data frame contains the labels and groups of the databases and tools of metabolic reconstruction}
\format{
\describe{
  \item{label}{the databases and tools of the metabolic reconstruction}
  \item{group}{the classfication of the labels}
  }
}
\usage{
toolsTypes
}
\description{
A data frame contains the labels and groups of the databases and tools of metabolic reconstruction
}
\keyword{datasets}
