% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizProcess.R
\name{vizProcess}
\alias{vizProcess}
\title{vizProcess}
\usage{
vizProcess(text, stepsMatrix, stepTypes, contentTypes)
}
\arguments{
\item{text}{The characters processed with the collapse = ' '.}

\item{stepsMatrix}{A data frame contains the marker words, threshold value, steps, group and other information about the metabolic reconstruction. The default file is in the data.}

\item{stepTypes}{A data frame contains the labels and groups of the metabolic reconstructions steps. The default file is in the data.}

\item{contentTypes}{A data frame contains the labels and groups of the metabolic reconstructions content The default file is in the data.}
}
\value{
The pictures that visualize the steps of the metabolic reconstruction.
}
\description{
Create a function can visualiaze the steps used in the metabolic reconstruction.
}
\examples{
\donttest{exam <- vizProcess(text, stepsMatrix, stepTypes,contentTypes)}
}
