% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizTransformation.R
\name{vizTransformation}
\alias{vizTransformation}
\title{vizTransformation}
\usage{
vizTransformation(text, conversionMatrix, stepTypes, conversionTypes)
}
\arguments{
\item{text}{The characters processed with the collapse = ' '.}

\item{conversionMatrix}{A data frame contains the marker words, threshold value, steps, group and other transformation information about the metabolic reconstruction. The default file is in the data.}

\item{stepTypes}{A data frame contains the labels and groups of the metabolic reconstructions steps. The default file is in the data.}

\item{conversionTypes}{A data frame contains the labels and groups of the metabolic reconstructions transformation. The default file is in the data.}
}
\value{
The pictures that visualize the transformation of the metabolic reconstruction.
}
\description{
Input the txt and output the visualization of the steps, transformation and databases and tools.
}
\examples{
\donttest{exam <- vizTransformation(text, conversionMatrix, stepTypes, conversionTypes)}
}
