% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordsMatrix}
\alias{wordsMatrix}
\title{A data frame produced by the function get_term_matrix}
\format{
A data frame contains 4 variables
\describe{
\item{freq}{the frequency of the word occurs in a article.}
\item{prevalent}{the words type that often occurs in a article.}
\item{longest}{the longest type of a word.}
\item{shortest}{the shortest type of a word.}
}
}
\usage{
wordsMatrix
}
\description{
A data frame produced by the function get_term_matrix
}
\keyword{datasets}
